Move demo modules into package
