+++ replenishment/algorithms/exponential_smoothing.py
