+++ replenishment/algorithms/moving_average.py
